Develop

+++ runonrouter.sh
@@ -4,8 +4,26 @@ set -e;
 ADDR="$1";
 SRC="$2";
 FILENAME=$(basename "$SRC");
-echo $FILENAME
 ARGS="${*:3}";
 
+echo "Usage: $(basename $0) <user>@<router.ip> <script.filename> [arg1 [arg2]]";
+echo "---";
+if [ $# -lt 2 ]; then
+  echo "At least two arguments mandatory.";
+  exit 1;
+fi;
+echo "Address: $ADDR";
+
+if [ "$SRC" = "" ] || [ ! -f "$SRC" ]; then
+  echo "File \"$SRC\" not found.";
+  exit 1;
+fi;
+
+echo "Script:  $SRC";
+echo "Args:    $ARGS";
+
+echo "Copying script to router..";
 scp "$SRC" "$ADDR:/tmp/$FILENAME";
-ssh "$ADDR" "chmod +x /tmp/$FILENAME && /tmp/$FILENAME $ARGS";
+
+echo "Running script on router..";
+ssh "$ADDR" "sh /tmp/$FILENAME $ARGS";
